+++ qa/restart-upgrade/src/test/java/org/opensearch/neuralsearch/bwc/restart/HybridSearchWithRescoreIT.java
@@ -55,7 +55,8 @@ public class HybridSearchWithRescoreIT extends AbstractRestartUpgradeRestTestCas
                 SEARCH_PIPELINE_NAME,
                 DEFAULT_NORMALIZATION_METHOD,
                 DEFAULT_COMBINATION_METHOD,
-                Map.of(PARAM_NAME_WEIGHTS, Arrays.toString(new float[] { 0.3f, 0.7f }))
+                Map.of(PARAM_NAME_WEIGHTS, Arrays.toString(new float[] { 0.3f, 0.7f })),
+                false
             );
         } else {
             String modelId = null;