+++ qa/rolling-upgrade/src/test/java/org/opensearch/neuralsearch/bwc/rolling/HybridSearchRelevancyIT.java
@@ -108,7 +108,8 @@ public class HybridSearchRelevancyIT extends AbstractRollingUpgradeTestCase {
                     SEARCH_PIPELINE_NAME,
                     "l2",
                     "arithmetic_mean",
-                    Map.of("weights", Arrays.toString(new float[] { 0.5f, 0.5f }))
+                    Map.of("weights", Arrays.toString(new float[] { 0.5f, 0.5f })),
+                    false
                 );
 
                 // execute hybrid query and store results