+++ src/test/java/org/opensearch/neuralsearch/processor/NormalizationProcessorWorkflowTests.java
@@ -506,9 +506,11 @@ public class NormalizationProcessorWorkflowTests extends OpenSearchTestCase {
         SearchPhaseContext searchPhaseContext = mock(SearchPhaseContext.class);
         when(searchPhaseContext.getNumShards()).thenReturn(4);
         SearchRequest searchRequest = mock(SearchRequest.class);
+        when(searchPhaseContext.getRequest()).thenReturn(searchRequest);
         SearchSourceBuilder searchSourceBuilder = new SearchSourceBuilder();
         searchSourceBuilder.from(17);
         when(searchPhaseContext.getRequest()).thenReturn(searchRequest);
+        when(searchPhaseContext.getRequest().source()).thenReturn(searchSourceBuilder);
 
         NormalizationProcessorWorkflowExecuteRequest normalizationExecuteDto = NormalizationProcessorWorkflowExecuteRequest.builder()
             .querySearchResults(querySearchResults)